+++ docs/changelog.md
@@ -20,6 +20,10 @@ Requires libvips v8.15.0
 * Options for `trim` operation must be an Object, add new `lineArt` option.
   [#2363](https://github.com/lovell/sharp/issues/2363)
 
+* Improve luminance of `tint` operation with weighting function.
+  [#3338](https://github.com/lovell/sharp/issues/3338)
+  [@jcupitt](https://github.com/jcupitt)
+
 * Ensure all `Error` objects contain a `stack` property.
   [#3653](https://github.com/lovell/sharp/issues/3653)