+++ docs/changelog.md
@@ -4,7 +4,7 @@
 
 Requires libvips v8.15.0
 
-### v0.33.0 - TBD
+### v0.33.0 - 29th November 2023
 
 * Drop support for Node.js 14 and 16, now requires Node.js >= 18.17.0